Move covid-specific information to covid flavor
