+++ rdt-open-reader/src/covidIndonesia/assets/ec_client_classification.json
@@ -1,0 +1,19 @@
+{
+  "case_classification_rules": [
+    {
+      "comment": "COVID-19 patient details",
+      "rule": {
+        "type": "event",
+        "fields": [
+          {
+            "field": "eventType",
+            "field_value": "covid_patient_registration",
+            "creates_case": [
+              "rdt_patients"
+            ]
+          }
+        ]
+      }
+    }
+  ]
+}
