+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/CovidLoginActivity.java
@@ -7,13 +7,15 @@ import com.vijay.jsonwizard.interfaces.FormWidgetFactory;
 import java.util.Map;
 
 import io.ona.rdt.interactor.RDTJsonFormInteractor;
+import io.ona.rdt.widget.CovidRDTLabelFactory;
 import io.ona.rdt.widget.GoogleCovidRDTBarcodeFactory;
 import io.ona.rdt.widget.OneScanCovidRDTBarcodeFactory;
 import io.ona.rdt.widget.UWCovidRDTCaptureFactory;
 
-import static io.ona.rdt.util.Constants.Widget.GOOGLE_COVID_BARCODE_READER;
-import static io.ona.rdt.util.Constants.Widget.ONE_SCAN_COVID_BARCODE_READER;
-import static io.ona.rdt.util.Constants.Widget.UW_COVID_RDT_CAPTURE;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.LABEL;
+import static io.ona.rdt.util.CovidConstants.Widget.GOOGLE_COVID_BARCODE_READER;
+import static io.ona.rdt.util.CovidConstants.Widget.ONE_SCAN_COVID_BARCODE_READER;
+import static io.ona.rdt.util.CovidConstants.Widget.UW_COVID_RDT_CAPTURE;
 
 /**
  * Created by Vincent Karuri on 19/06/2020
@@ -31,6 +33,7 @@ public class CovidLoginActivity extends LoginActivity {
         widgetFactoryMap.put(UW_COVID_RDT_CAPTURE, new UWCovidRDTCaptureFactory());
         widgetFactoryMap.put(GOOGLE_COVID_BARCODE_READER, new GoogleCovidRDTBarcodeFactory());
         widgetFactoryMap.put(ONE_SCAN_COVID_BARCODE_READER, new OneScanCovidRDTBarcodeFactory());
+        widgetFactoryMap.put(LABEL, new CovidRDTLabelFactory());
     }
 
     @Override
