+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/CovidPatientRegisterActivity.java
@@ -7,7 +7,7 @@ import org.smartregister.view.fragment.BaseRegisterFragment;
 import io.ona.rdt.R;
 import io.ona.rdt.fragment.CovidPatientRegisterFragment;
 
-import static io.ona.rdt.util.Constants.Form.SAMPLE_DELIVERY_DETAILS_FORM;
+import static io.ona.rdt.util.CovidConstants.Form.SAMPLE_DELIVERY_DETAILS_FORM;
 
 /**
  * Created by Vincent Karuri on 16/06/2020
