+++ rdt-open-reader/src/main/java/io/ona/rdt/interactor/PatientRegisterFragmentInteractor.java
@@ -33,15 +33,15 @@ import io.ona.rdt.util.FormLauncher;
 import timber.log.Timber;
 
 import static com.vijay.jsonwizard.utils.Utils.showToast;
-import static io.ona.rdt.util.Constants.Encounter.COVID_PATIENT_REGISTRATION;
-import static io.ona.rdt.util.Constants.Encounter.COVID_RDT_TEST;
-import static io.ona.rdt.util.Constants.Encounter.PATIENT_DIAGNOSTICS;
+import static io.ona.rdt.util.CovidConstants.Encounter.COVID_PATIENT_REGISTRATION;
+import static io.ona.rdt.util.CovidConstants.Encounter.COVID_RDT_TEST;
+import static io.ona.rdt.util.CovidConstants.Encounter.PATIENT_DIAGNOSTICS;
 import static io.ona.rdt.util.Constants.Encounter.PATIENT_REGISTRATION;
 import static io.ona.rdt.util.Constants.Encounter.PCR_RESULT;
 import static io.ona.rdt.util.Constants.Encounter.RDT_TEST;
-import static io.ona.rdt.util.Constants.Encounter.SAMPLE_COLLECTION;
-import static io.ona.rdt.util.Constants.Encounter.SAMPLE_DELIVERY_DETAILS;
-import static io.ona.rdt.util.Constants.Encounter.SUPPORT_INVESTIGATION;
+import static io.ona.rdt.util.CovidConstants.Encounter.SAMPLE_COLLECTION;
+import static io.ona.rdt.util.CovidConstants.Encounter.SAMPLE_DELIVERY_DETAILS;
+import static io.ona.rdt.util.CovidConstants.Encounter.SUPPORT_INVESTIGATION;
 import static io.ona.rdt.util.Constants.FormFields.DETAILS;
 import static io.ona.rdt.util.Constants.FormFields.DOB;
 import static io.ona.rdt.util.Constants.FormFields.ENCOUNTER_TYPE;
@@ -49,16 +49,16 @@ import static io.ona.rdt.util.Constants.FormFields.ENTITY_ID;
 import static io.ona.rdt.util.Constants.FormFields.METADATA;
 import static io.ona.rdt.util.Constants.FormFields.PATIENT_AGE;
 import static io.ona.rdt.util.Constants.FormFields.RDT_ID;
-import static io.ona.rdt.util.Constants.FormFields.COVID_SAMPLE_ID;
-import static io.ona.rdt.util.Constants.Table.COVID_PATIENTS;
-import static io.ona.rdt.util.Constants.Table.COVID_RDT_TESTS;
-import static io.ona.rdt.util.Constants.Table.PATIENT_DIAGNOSTIC_RESULTS;
+import static io.ona.rdt.util.CovidConstants.FormFields.COVID_SAMPLE_ID;
+import static io.ona.rdt.util.CovidConstants.Table.COVID_PATIENTS;
+import static io.ona.rdt.util.CovidConstants.Table.COVID_RDT_TESTS;
+import static io.ona.rdt.util.CovidConstants.Table.PATIENT_DIAGNOSTIC_RESULTS;
 import static io.ona.rdt.util.Constants.Table.PCR_RESULTS;
 import static io.ona.rdt.util.Constants.Table.RDT_PATIENTS;
 import static io.ona.rdt.util.Constants.Table.RDT_TESTS;
-import static io.ona.rdt.util.Constants.Table.SAMPLE_COLLECTIONS;
-import static io.ona.rdt.util.Constants.Table.SAMPLE_DELIVERY_RECORDS;
-import static io.ona.rdt.util.Constants.Table.SUPPORT_INVESTIGATIONS;
+import static io.ona.rdt.util.CovidConstants.Table.SAMPLE_COLLECTIONS;
+import static io.ona.rdt.util.CovidConstants.Table.SAMPLE_DELIVERY_RECORDS;
+import static io.ona.rdt.util.CovidConstants.Table.SUPPORT_INVESTIGATIONS;
 import static io.ona.rdt.util.RDTJsonFormUtils.getFormTag;
 import static io.ona.rdt.util.Utils.isCovidApp;
 import static org.smartregister.util.JsonFormUtils.KEY;
