+++ rdt-open-reader/src/main/java/io/ona/rdt/util/Constants.java
@@ -7,12 +7,6 @@ import io.ona.rdt.BuildConfig;
  */
 public interface Constants {
 
-    interface Widget {
-        String UW_COVID_RDT_CAPTURE = "uw_covid_rdt_capture";
-        String GOOGLE_COVID_BARCODE_READER = "google_covid_barcode_reader";
-        String ONE_SCAN_COVID_BARCODE_READER = "one_scan_covid_barcode_reader";
-    }
-
     interface Format {
         String BULLET_DOT = " \u00B7 ";
         String PROFILE_DATE_FORMAT = "dd MMM yyyy";
@@ -32,6 +26,7 @@ public interface Constants {
     interface RDTType {
         String ONA_RDT = "malaria-experimental";
         String CARESTART_RDT = "malaria-carestart";
+        String RDT_TYPE = "rdt_type";
     }
 
     interface FormFields {
@@ -47,26 +42,10 @@ public interface Constants {
         String MANUAL_EXPIRATION_DATE = "manual_expiration_date";
         String LBL_CARE_START = "lbl_care_start";
         String LBL_SCAN_QR_CODE = "lbl_scan_qr_code";
-        String LBL_SCAN_BARCODE = "lbl_scan_barcode";
-        String LBL_ENTER_RDT_MANUALLY = "lbl_enter_rdt_manually";
-        String LBL_CONDUCT_RDT = "lbl_conduct_rdt";
-        String LBL_SKIP_RDT_TEST = "lbl_skip_rdt_test";
-        String LBL_COLLECT_RESPIRATORY_SAMPLE = "lbl_collect_respiratory_sample";
-        String LBL_SKIP_RESPIRATORY_SAMPLE_COLLECTION = "lbl_skip_respiratory_sample_collection";
-        String LBL_SCAN_RESPIRATORY_SPECIMEN_BARCODE = "lbl_scan_respiratory_specimen_barcode";
-        String LBL_AFFIX_RESPIRATORY_SPECIMEN_LABEL = "lbl_affix_respiratory_specimen_label";
         String LBL_RDT_ID = "lbl_rdt_id";
         String LBL_PATIENT_NAME = "lbl_patient_name";
         String LBL_PATIENT_GENDER_AND_ID = "lbl_patient_gender_and_id";
-        String LBL_RESPIRATORY_SAMPLE_ID = "lbl_respiratory_sample_id";
-        String LBL_ADD_XRAY_RESULTS = "lbl_add_xray_results";
-        String LBL_SKIP_XRAY_RESULTS = "lbl_skip_xray_results";
-        String LBL_ADD_WBC_RESULTS = "lbl_add_wbc_results";
-        String LBL_SKIP_WBC_RESULTS = "lbl_skip_wbc_results";
-        String LBL_SCAN_SAMPLE_BARCODE = "lbl_scan_sample_barcode";
-        String LBL_ENTER_SAMPLE_DETAILS_MANUALLY = "lbl_enter_sample_details_manually";
         String RDT_ID = "rdt_id";
-        String COVID_SAMPLE_ID = "covid_sample_id";
         String RDT_CAPTURE_TOP_LINE_RESULT = "rdt_capture_top_line_result";
         String RDT_CAPTURE_MIDDLE_LINE_RESULT = "rdt_capture_middle_line_result";
         String RDT_CAPTURE_BOTTOM_LINE_RESULT = "rdt_capture_bottom_line_result";
@@ -103,11 +82,6 @@ public interface Constants {
     interface Form {
         String PATIENT_REGISTRATION_FORM = "json.form-" + BuildConfig.LOCALE + "/patient-registration-form.json";
         String RDT_TEST_FORM = "json.form-" + BuildConfig.LOCALE + "/rdt-capture-form.json";
-        String PATIENT_DIAGNOSTICS_FORM = "json.form-" + BuildConfig.LOCALE + "/patient-diagnostics-form.json";
-        String SAMPLE_COLLECTION_FORM = "json.form-" + BuildConfig.LOCALE + "/sample-collection-form.json";
-        String SAMPLE_DELIVERY_DETAILS_FORM = "json.form-" + BuildConfig.LOCALE + "/sample-delivery-details-form.json";
-        String SUPPORT_INVESTIGATION_FORM = "json.form-" + BuildConfig.LOCALE + "/support-investigation-form.json";
-        String RDT_TYPE = "rdt_type";
     }
 
     interface Test {
@@ -144,43 +118,18 @@ public interface Constants {
         String SCAN_CARESTART_PAGE = "scan_carestart_page";
         String TWENTY_MIN_COUNTDOWN_TIMER_PAGE = "twenty_min_countdown_timer_page";
         String TAKE_IMAGE_OF_RDT_PAGE = "take_image_of_rdt_page";
-        String COVID_SCAN_BARCODE_PAGE = "covid_scan_barcode_page";
-        String COVID_CONDUCT_RDT_PAGE = "covid_conduct_rdt_page";
-        String COVID_RESPIRATORY_SPECIMEN_COLLECTION_OPT_IN_PAGE = "covid_respiratory_specimen_collection_opt_in_page";
-        String COVID_COLLECT_RESPIRATORY_SPECIMEN_PAGE = "covid_collect_respiratory_specimen_page";
-        String COVID_TEST_COMPLETE_PAGE = "covid_test_complete_page";
-        String COVID_AFFIX_RESPIRATORY_SAMPLE_ID_PAGE = "covid_affix_respiratory_sample_id_page";
-        String COVID_ONE_SCAN_WIDGET_SPECIMEN_PAGE  = "covid_one_scan_widget_specimen_page";
-        String COVID_XRAY_PAGE = "covid_xray_page";
-        String COVID_OPT_IN_WBC_PAGE = "covid_opt_in_wbc_page";
-        String COVID_WBC_PAGE = "covid_wbc_page";
-        String COVID_SUPPORT_INVESTIGATION_COMPLETE_PAGE = "covid_support_investigation_complete_page";
-        String COVID_SCAN_SAMPLE_FOR_DELIVERY_PAGE = "covid_scan_sample_for_delivery_page";
-        String COVID_ENTER_DELIVERY_DETAILS_PAGE = "covid_enter_delivery_details_page";
     }
 
     interface Encounter {
         String PATIENT_REGISTRATION = "patient_registration";
-        String COVID_PATIENT_REGISTRATION = "covid_patient_registration";
         String RDT_TEST = "rdt_test";
-        String COVID_RDT_TEST = "covid_rdt_test";
         String PCR_RESULT = "pcr_result";
-        String PATIENT_DIAGNOSTICS = "patient_diagnostics";
-        String SAMPLE_COLLECTION = "sample_collection";
-        String SAMPLE_DELIVERY_DETAILS = "sample_delivery_details";
-        String SUPPORT_INVESTIGATION = "support_investigation";
     }
 
     interface Table {
         String RDT_PATIENTS = "rdt_patients";
-        String COVID_PATIENTS = "covid_patients";
         String RDT_TESTS = "rdt_tests";
-        String COVID_RDT_TESTS = "covid_rdt_tests";
         String PCR_RESULTS = "pcr_results";
         String MICROSCOPY_RESULTS = "microscopy_results";
-        String PATIENT_DIAGNOSTIC_RESULTS = "patient_diagnostic_results";
-        String SAMPLE_COLLECTIONS = "sample_collections";
-        String SAMPLE_DELIVERY_RECORDS = "sample_delivery_records";
-        String SUPPORT_INVESTIGATIONS = "support_investigations";
     }
 }
