+++ rdt-open-reader/src/test/java/io/ona/rdt/interactor/PatientRegisterFragmentInteractorTest.java
@@ -47,36 +47,35 @@ import io.ona.rdt.util.StepStateConfig;
 
 import static com.vijay.jsonwizard.constants.JsonFormConstants.KEY;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.VALUE;
-import static io.ona.rdt.util.Constants.Encounter.COVID_PATIENT_REGISTRATION;
-import static io.ona.rdt.util.Constants.Encounter.COVID_RDT_TEST;
-import static io.ona.rdt.util.Constants.Encounter.PATIENT_DIAGNOSTICS;
+import static io.ona.rdt.util.CovidConstants.Encounter.COVID_PATIENT_REGISTRATION;
+import static io.ona.rdt.util.CovidConstants.Encounter.COVID_RDT_TEST;
+import static io.ona.rdt.util.CovidConstants.Encounter.PATIENT_DIAGNOSTICS;
 import static io.ona.rdt.util.Constants.Encounter.PATIENT_REGISTRATION;
 import static io.ona.rdt.util.Constants.Encounter.PCR_RESULT;
 import static io.ona.rdt.util.Constants.Encounter.RDT_TEST;
-import static io.ona.rdt.util.Constants.Encounter.SAMPLE_COLLECTION;
-import static io.ona.rdt.util.Constants.Encounter.SAMPLE_DELIVERY_DETAILS;
-import static io.ona.rdt.util.Constants.Encounter.SUPPORT_INVESTIGATION;
+import static io.ona.rdt.util.CovidConstants.Encounter.SAMPLE_COLLECTION;
+import static io.ona.rdt.util.CovidConstants.Encounter.SAMPLE_DELIVERY_DETAILS;
+import static io.ona.rdt.util.CovidConstants.Encounter.SUPPORT_INVESTIGATION;
 import static io.ona.rdt.util.Constants.Form.RDT_TEST_FORM;
-import static io.ona.rdt.util.Constants.Form.SAMPLE_COLLECTION_FORM;
+import static io.ona.rdt.util.CovidConstants.Form.SAMPLE_COLLECTION_FORM;
 import static io.ona.rdt.util.Constants.FormFields.ENCOUNTER_TYPE;
 import static io.ona.rdt.util.Constants.FormFields.RDT_ID;
-import static io.ona.rdt.util.Constants.FormFields.COVID_SAMPLE_ID;
+import static io.ona.rdt.util.CovidConstants.FormFields.COVID_SAMPLE_ID;
 import static io.ona.rdt.util.Constants.RequestCodes.REQUEST_CODE_GET_JSON;
 import static io.ona.rdt.util.Constants.Step.RDT_ID_KEY;
-import static io.ona.rdt.util.Constants.Table.COVID_PATIENTS;
-import static io.ona.rdt.util.Constants.Table.COVID_RDT_TESTS;
-import static io.ona.rdt.util.Constants.Table.PATIENT_DIAGNOSTIC_RESULTS;
+import static io.ona.rdt.util.CovidConstants.Table.COVID_PATIENTS;
+import static io.ona.rdt.util.CovidConstants.Table.COVID_RDT_TESTS;
+import static io.ona.rdt.util.CovidConstants.Table.PATIENT_DIAGNOSTIC_RESULTS;
 import static io.ona.rdt.util.Constants.Table.PCR_RESULTS;
 import static io.ona.rdt.util.Constants.Table.RDT_PATIENTS;
 import static io.ona.rdt.util.Constants.Table.RDT_TESTS;
-import static io.ona.rdt.util.Constants.Table.SAMPLE_COLLECTIONS;
-import static io.ona.rdt.util.Constants.Table.SAMPLE_DELIVERY_RECORDS;
-import static io.ona.rdt.util.Constants.Table.SUPPORT_INVESTIGATIONS;
+import static io.ona.rdt.util.CovidConstants.Table.SAMPLE_COLLECTIONS;
+import static io.ona.rdt.util.CovidConstants.Table.SAMPLE_DELIVERY_RECORDS;
+import static io.ona.rdt.util.CovidConstants.Table.SUPPORT_INVESTIGATIONS;
 import static io.ona.rdt.util.Utils.isCovidApp;
 import static io.ona.rdt.util.Utils.isMalariaApp;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
