+++ rdt-open-reader/src/test/java/io/ona/rdt/widget/CustomRDTCaptureFactoryTest.java
@@ -36,7 +36,7 @@ import static com.vijay.jsonwizard.constants.JsonFormConstants.RDT_CAPTURE_CODE;
 import static io.ona.rdt.util.Constants.FormFields.RDT_CAPTURE_TOP_LINE_RESULT;
 import static io.ona.rdt.util.Constants.FormFields.RDT_CAPTURE_BOTTOM_LINE_RESULT;
 import static io.ona.rdt.util.Constants.FormFields.RDT_CAPTURE_MIDDLE_LINE_RESULT;
-import static io.ona.rdt.util.Constants.Form.RDT_TYPE;
+import static io.ona.rdt.util.Constants.RDTType.RDT_TYPE;
 import static io.ona.rdt.util.Constants.Test.CASSETTE_BOUNDARY;
 import static io.ona.rdt.util.Constants.Test.CROPPED_IMG_ID;
 import static io.ona.rdt.util.Constants.Test.FLASH_ON;
