+++ rdt-open-reader/src/test/java/io/ona/rdt/widget/RDTLabelFactoryTest.java
@@ -30,23 +30,24 @@ import io.ona.rdt.callback.OnLabelClickedListener;
 import io.ona.rdt.fragment.RDTJsonFormFragment;
 import io.ona.rdt.presenter.RDTJsonFormFragmentPresenter;
 import io.ona.rdt.util.Constants;
+import io.ona.rdt.util.CovidConstants;
 import io.ona.rdt.util.StepStateConfig;
 
 import static com.vijay.jsonwizard.constants.JsonFormConstants.KEY;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.NEXT;
-import static io.ona.rdt.util.Constants.Step.COVID_AFFIX_RESPIRATORY_SAMPLE_ID_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_COLLECT_RESPIRATORY_SPECIMEN_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_CONDUCT_RDT_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_ENTER_DELIVERY_DETAILS_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_ONE_SCAN_WIDGET_SPECIMEN_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_OPT_IN_WBC_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_RESPIRATORY_SPECIMEN_COLLECTION_OPT_IN_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_SCAN_BARCODE_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_SCAN_SAMPLE_FOR_DELIVERY_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_SUPPORT_INVESTIGATION_COMPLETE_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_TEST_COMPLETE_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_WBC_PAGE;
-import static io.ona.rdt.util.Constants.Step.COVID_XRAY_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_AFFIX_RESPIRATORY_SAMPLE_ID_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_COLLECT_RESPIRATORY_SPECIMEN_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_CONDUCT_RDT_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_ENTER_DELIVERY_DETAILS_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_ONE_SCAN_WIDGET_SPECIMEN_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_OPT_IN_WBC_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_RESPIRATORY_SPECIMEN_COLLECTION_OPT_IN_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_SCAN_BARCODE_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_SCAN_SAMPLE_FOR_DELIVERY_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_SUPPORT_INVESTIGATION_COMPLETE_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_TEST_COMPLETE_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_WBC_PAGE;
+import static io.ona.rdt.util.CovidConstants.Step.COVID_XRAY_PAGE;
 import static io.ona.rdt.util.Constants.Step.MANUAL_EXPIRATION_DATE_ENTRY_PAGE;
 import static io.ona.rdt.util.Constants.Step.SCAN_CARESTART_PAGE;
 import static io.ona.rdt.util.Constants.Step.SCAN_QR_PAGE;
@@ -128,59 +129,43 @@ public class RDTLabelFactoryTest extends PowerMockTest {
         verify(jsonFormFragment.getRdtActivity()).setRdtType(eq(Constants.RDTType.ONA_RDT));
         verify(presenter).moveToNextStep(eq(SCAN_QR_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_SCAN_BARCODE);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_SCAN_BARCODE);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_SCAN_BARCODE_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_ENTER_RDT_MANUALLY);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_ENTER_RDT_MANUALLY);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(MANUAL_EXPIRATION_DATE_ENTRY_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_CONDUCT_RDT);
-        onLabelClickedListener.onClick(mock(View.class));
-        verify(presenter).moveToNextStep(eq(COVID_CONDUCT_RDT_PAGE));
-
-        jsonObject.put(KEY, Constants.FormFields.LBL_SKIP_RDT_TEST);
-        onLabelClickedListener.onClick(mock(View.class));
-        verify(presenter).moveToNextStep(eq(COVID_RESPIRATORY_SPECIMEN_COLLECTION_OPT_IN_PAGE));
-
-        jsonObject.put(KEY, Constants.FormFields.LBL_COLLECT_RESPIRATORY_SAMPLE);
-        onLabelClickedListener.onClick(mock(View.class));
-        verify(presenter).moveToNextStep(eq(COVID_COLLECT_RESPIRATORY_SPECIMEN_PAGE));
-
-        jsonObject.put(KEY, Constants.FormFields.LBL_SKIP_RESPIRATORY_SAMPLE_COLLECTION);
-        onLabelClickedListener.onClick(mock(View.class));
-        verify(presenter).moveToNextStep(eq(COVID_TEST_COMPLETE_PAGE));
-
-        jsonObject.put(KEY, Constants.FormFields.LBL_SCAN_RESPIRATORY_SPECIMEN_BARCODE);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_SCAN_RESPIRATORY_SPECIMEN_BARCODE);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_ONE_SCAN_WIDGET_SPECIMEN_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_AFFIX_RESPIRATORY_SPECIMEN_LABEL);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_AFFIX_RESPIRATORY_SPECIMEN_LABEL);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_AFFIX_RESPIRATORY_SAMPLE_ID_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_ADD_XRAY_RESULTS);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_ADD_XRAY_RESULTS);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_XRAY_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_SKIP_XRAY_RESULTS);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_SKIP_XRAY_RESULTS);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_OPT_IN_WBC_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_ADD_WBC_RESULTS);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_ADD_WBC_RESULTS);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_WBC_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_SKIP_WBC_RESULTS);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_SKIP_WBC_RESULTS);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_SUPPORT_INVESTIGATION_COMPLETE_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_SCAN_SAMPLE_BARCODE);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_SCAN_SAMPLE_BARCODE);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_SCAN_SAMPLE_FOR_DELIVERY_PAGE));
 
-        jsonObject.put(KEY, Constants.FormFields.LBL_ENTER_SAMPLE_DETAILS_MANUALLY);
+        jsonObject.put(KEY, CovidConstants.FormFields.LBL_ENTER_SAMPLE_DETAILS_MANUALLY);
         onLabelClickedListener.onClick(mock(View.class));
         verify(presenter).moveToNextStep(eq(COVID_ENTER_DELIVERY_DETAILS_PAGE));
     }
